# docs/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "bdzbridge",
  "version": "0.1.0"
 },
 "paths": {
...
  "/api/v1/rules": {
   "get": {
    "tags": [
     "rules"
    ],
    "summary": "Rules",
    "description": "The keyword auto-reservation rules.",
    "operationId": "rules_api_v1_rules_get",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/Rule"
         },
         "title": "Response Rules Api V1 Rules Get"
        }
       }
      }
     }
    }
   },
   "post": {
    "tags": [
     "rules"
    ],
    "summary": "Rule Create",
    "operationId": "rule_create_api_v1_rules_post",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "run",
      "in": "query",
      "required": false,
      "schema": {
       "type": "boolean",
       "description": "apply every rule right away (reserves on the recorder)",
       "default": false,
       "title": "Run"
      },
      "description": "apply every rule right away (reserves on the recorder)"
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/RuleCreate"
       }
      }
     }
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Rule"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/rules/log": {
   "get": {
    "tags": [
     "rules"
    ],
    "summary": "Rules Log",
    "operationId": "rules_log_api_v1_rules_log_get",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "limit",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 500,
       "default": 50,
       "title": "Limit"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/AutoLogEntry"
         },
         "title": "Response Rules Log Api V1 Rules Log Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/rules/run": {
   "post": {
    "tags": [
     "rules"
    ],
    "summary": "Rules Run",
    "description": "Apply every enabled rule now (they also run after each guide refresh); reports what was reserved.",
    "operationId": "rules_run_api_v1_rules_run_post",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/AutoRunResult"
        }
       }
      }
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ]
   }
  },
  "/api/v1/rules/{rule_id}/matches": {
   "get": {
    "tags": [
     "rules"
    ],
    "summary": "Rule Matches",
    "description": "Upcoming programs the rule matches (reserved or not).",
    "operationId": "rule_matches_api_v1_rules__rule_id__matches_get",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "rule_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Rule Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/Program"
         },
         "title": "Response Rule Matches Api V1 Rules  Rule Id  Matches Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/rules/{rule_id}": {
   "patch": {
    "tags": [
     "rules"
    ],
    "summary": "Rule Update",
    "description": "Enable or disable a rule, or change its quality or title-only matching.",
    "operationId": "rule_update_api_v1_rules__rule_id__patch",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "rule_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Rule Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/RuleUpdate"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Rule"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "delete": {
    "tags": [
     "rules"
    ],
    "summary": "Rule Delete",
    "description": "Delete a rule and its log.",
    "operationId": "rule_delete_api_v1_rules__rule_id__delete",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "rule_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Rule Id"
      }
     }
    ],
    "responses": {
     "204": {
      "description": "Successful Response"
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/recorder-rules": {
   "get": {
    "tags": [
     "rules"
    ],
    "summary": "Recorder Rules",
    "description": "The keyword conditions held by the recorder itself (おまかせ・まる録). These record without this server. The channel narrowing set on the recorder's screen is not reported.",
    "operationId": "recorder_rules_api_v1_recorder_rules_get",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "items": {
          "$ref": "#/components/schemas/RecorderRule"
         },
         "type": "array",
         "title": "Response Recorder Rules Api V1 Recorder Rules Get"
        }
       }
      }
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ]
   },
   "post": {
    "tags": [
     "rules"
    ],
    "summary": "Recorder Rule Create",
    "description": "Register a condition on the recorder itself. The recorder composes the name; the channel cannot be set this way.",
    "operationId": "recorder_rule_create_api_v1_recorder_rules_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/RecorderRuleCreate"
       }
      }
     },
     "required": true
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/RecorderRule"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ]
   }
  },
  "/api/v1/recorder-rules/{rule_id}": {
   "delete": {
    "tags": [
     "rules"
    ],
    "summary": "Recorder Rule Delete",
    "description": "Remove a condition from the recorder, whoever made it. Ids change whenever the recorder's screen edits a condition, so read the list first.",
    "operationId": "recorder_rule_delete_api_v1_recorder_rules__rule_id__delete",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "rule_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Rule Id"
      }
     }
    ],
    "responses": {
     "204": {
      "description": "Successful Response"
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/monitor/run": {
   "post": {
    "tags": [
     "rules"
    ],
    "summary": "Monitor Run",
    "description": "Check free space and conflicting reservations now (normally runs after every EPG refresh).",
    "operationId": "monitor_run_api_v1_monitor_run_post",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/MonitorResult"
        }
       }
      }
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ]
   }
  },
  "/api/v1/notify": {
   "get": {
    "tags": [
     "rules"
    ],
    "summary": "Notify Status",
    "description": "Which notification channels are configured (SMTP, webhook) and the free-space warning threshold.",
    "operationId": "notify_status_api_v1_notify_get",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/NotifyStatus"
        }
       }
      }
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ]
   }
  },
  "/api/v1/notify/test": {
   "post": {
    "tags": [
     "rules"
    ],
    "summary": "Notify Test",
    "description": "Send a test message through every configured channel.",
    "operationId": "notify_test_api_v1_notify_test_post",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/NotifyStatus"
        }
       }
      }
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ]
   }
  },
...
 },
 "components": {
  "schemas": {
   "AutoLogEntry": {
    "properties": {
     "id": {
      "type": "integer",
      "title": "Id"
     },
     "rule_id": {
      "type": "integer",
      "title": "Rule Id"
     },
     "rule_query": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Rule Query"
     },
     "broadcasting": {
      "type": "string",
      "title": "Broadcasting"
     },
     "service_id": {
      "type": "integer",
      "title": "Service Id"
     },
     "event_id": {
      "type": "integer",
      "title": "Event Id"
     },
     "title": {
      "type": "string",
      "title": "Title"
     },
     "start": {
      "type": "string",
      "format": "date-time",
      "title": "Start"
     },
     "status": {
      "type": "string",
      "enum": [
       "reserved",
       "conflict",
       "error"
      ],
      "title": "Status"
     },
     "message": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Message"
     },
     "at": {
      "type": "string",
      "format": "date-time",
      "title": "At"
     }
    },
    "type": "object",
    "required": [
     "id",
     "rule_id",
     "broadcasting",
     "service_id",
     "event_id",
     "title",
     "start",
     "status",
     "at"
    ],
    "title": "AutoLogEntry"
   },
   "AutoRunResult": {
    "properties": {
     "rules": {
      "type": "integer",
      "title": "Rules"
     },
     "checked": {
      "type": "integer",
      "title": "Checked"
     },
     "reserved": {
      "type": "integer",
      "title": "Reserved"
     },
     "conflicts": {
      "type": "integer",
      "title": "Conflicts"
     },
     "errors": {
      "type": "integer",
      "title": "Errors"
     },
     "notified": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "Notified",
      "description": "channels that delivered the report: email, webhook"
     },
     "at": {
      "anyOf": [
       {
        "type": "string",
        "format": "date-time"
       },
       {
        "type": "null"
       }
      ],
      "title": "At"
     }
    },
    "type": "object",
    "required": [
     "rules",
     "checked",
     "reserved",
     "conflicts",
     "errors"
    ],
    "title": "AutoRunResult"
   },
...
   "Genre": {
    "properties": {
     "level1": {
      "type": "integer",
      "title": "Level1"
     },
     "level2": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Level2",
      "description": "None stands for the whole level-1 genre, as a recorder condition can"
     },
     "label": {
      "type": "string",
      "title": "Label"
     }
    },
    "type": "object",
    "required": [
     "level1",
     "level2",
     "label"
    ],
    "title": "Genre"
   },
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "MonitorResult": {
    "properties": {
     "free_gb": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Free Gb"
     },
     "low_space": {
      "type": "boolean",
      "title": "Low Space"
     },
     "new_conflicts": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "New Conflicts"
     },
     "notified": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "Notified"
     }
    },
    "type": "object",
    "required": [
     "low_space",
     "new_conflicts",
     "notified"
    ],
    "title": "MonitorResult"
   },
   "NotifyStatus": {
    "properties": {
     "configured": {
      "type": "boolean",
      "title": "Configured"
     },
     "email": {
      "type": "boolean",
      "title": "Email"
     },
     "webhook": {
      "type": "boolean",
      "title": "Webhook"
     },
     "to": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "To"
     },
     "free_gb": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Free Gb",
      "description": "low-space warning threshold, 0 = off"
     },
     "sent": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "Sent"
     }
    },
    "type": "object",
    "required": [
     "configured",
     "email",
     "webhook"
    ],
    "title": "NotifyStatus"
   },
...
   "Program": {
    "properties": {
     "broadcasting": {
      "type": "string",
      "enum": [
       "td",
       "bs",
       "cs",
       "bs4k",
       "cs4k"
      ],
      "title": "Broadcasting"
     },
     "service_id": {
      "type": "integer",
      "title": "Service Id"
     },
     "service_name": {
      "type": "string",
      "title": "Service Name"
     },
     "event_id": {
      "type": "integer",
      "title": "Event Id"
     },
     "start": {
      "type": "string",
      "format": "date-time",
      "title": "Start"
     },
     "end": {
      "type": "string",
      "format": "date-time",
      "title": "End"
     },
     "duration_sec": {
      "type": "integer",
      "title": "Duration Sec"
     },
     "title": {
      "type": "string",
      "title": "Title"
     },
     "description": {
      "type": "string",
      "title": "Description"
     },
     "extended": {
      "type": "string",
      "title": "Extended",
      "default": ""
     },
     "genres": {
      "items": {
       "$ref": "#/components/schemas/Genre"
      },
      "type": "array",
      "title": "Genres"
     },
     "copy_control": {
      "type": "integer",
      "title": "Copy Control"
     },
     "parental_rating": {
      "type": "integer",
      "title": "Parental Rating"
     },
     "is_reference": {
      "type": "boolean",
      "title": "Is Reference",
      "default": false
     },
     "ref_service_id": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Ref Service Id"
     },
     "ref_event_id": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Ref Event Id"
     }
    },
    "type": "object",
    "required": [
     "broadcasting",
     "service_id",
     "service_name",
     "event_id",
     "start",
     "end",
     "duration_sec",
     "title",
     "description",
     "genres",
     "copy_control",
     "parental_rating"
    ],
    "title": "Program"
   },
...
   "RecorderRule": {
    "properties": {
     "id": {
      "type": "string",
      "title": "Id"
     },
     "name": {
      "type": "string",
      "title": "Name",
      "description": "composed by the recorder from the genre and the keywords"
     },
     "keywords": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "Keywords"
     },
     "excluded": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "Excluded"
     },
     "logic": {
      "type": "string",
      "title": "Logic"
     },
     "logic_label": {
      "type": "string",
      "title": "Logic Label"
     },
     "genres": {
      "items": {
       "$ref": "#/components/schemas/Genre"
      },
      "type": "array",
      "title": "Genres"
     },
     "time_scope": {
      "type": "string",
      "title": "Time Scope"
     },
     "time_scope_label": {
      "type": "string",
      "title": "Time Scope Label"
     },
     "broadcasting_scope": {
      "type": "string",
      "title": "Broadcasting Scope"
     },
     "broadcasting_scope_label": {
      "type": "string",
      "title": "Broadcasting Scope Label"
     },
     "quality": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Quality",
      "description": "録画モード(地上/BS/CS)"
     },
     "quality_4k": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Quality 4K",
      "description": "録画モード(BS4K/CS4K), filled in by the recorder"
     },
     "destination": {
      "type": "string",
      "title": "Destination"
     }
    },
    "type": "object",
    "required": [
     "id",
     "name",
     "keywords",
     "excluded",
     "logic",
     "logic_label",
     "genres",
     "time_scope",
     "time_scope_label",
     "broadcasting_scope",
     "broadcasting_scope_label",
     "quality",
     "quality_4k",
     "destination"
    ],
    "title": "RecorderRule"
   },
   "RecorderRuleCreate": {
    "properties": {
     "keywords": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "maxItems": 5,
      "title": "Keywords",
      "description": "as the recorder's own screen allows: up to 5; a genre alone is also a condition"
     },
     "excluded": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "maxItems": 2,
      "title": "Excluded",
      "description": "up to 2"
     },
     "logic": {
      "type": "string",
      "enum": [
       "OR",
       "AND"
      ],
      "title": "Logic",
      "default": "OR"
     },
     "genre_level1": {
      "anyOf": [
       {
        "type": "integer",
        "maximum": 15.0,
        "minimum": 0.0
       },
       {
        "type": "null"
       }
      ],
      "title": "Genre Level1",
      "description": "ARIB level-1 genre; alone it means the whole genre"
     },
     "genre_level2": {
      "anyOf": [
       {
        "type": "integer",
        "maximum": 15.0,
        "minimum": 0.0
       },
       {
        "type": "null"
       }
      ],
      "title": "Genre Level2",
      "description": "the sub-genre within level1"
     },
     "time_scope": {
      "type": "string",
      "maxLength": 16,
      "title": "Time Scope",
      "description": "ALL or NIGHT are known; others are passed through",
      "default": "ALL"
     },
     "broadcasting_scope": {
      "type": "string",
      "maxLength": 16,
      "title": "Broadcasting Scope",
      "description": "ALL or TRD are known; others are passed through",
      "default": "ALL"
     },
     "quality": {
      "anyOf": [
       {
        "type": "string",
        "enum": [
         "DR",
         "XR",
         "XSR",
         "SR",
         "LSR",
         "LR",
         "ER",
         "EER"
        ]
       },
       {
        "type": "null"
       }
      ],
      "title": "Quality"
     }
    },
    "type": "object",
    "title": "RecorderRuleCreate",
    "description": "A condition for the recorder's own おまかせ・まる録, which then records by it without this server. The\nchannel narrowing the recorder's screen offers cannot be set over the LAN."
   },
...
   "Rule": {
    "properties": {
     "id": {
      "type": "integer",
      "title": "Id"
     },
     "query": {
      "type": "string",
      "title": "Query"
     },
     "broadcasting": {
      "anyOf": [
       {
        "type": "string",
        "enum": [
         "td",
         "bs",
         "cs",
         "bs4k",
         "cs4k"
        ]
       },
       {
        "type": "null"
       }
      ],
      "title": "Broadcasting"
     },
     "service_id": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Service Id"
     },
     "service_name": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Service Name"
     },
     "title_only": {
      "type": "boolean",
      "title": "Title Only"
     },
     "quality": {
      "type": "string",
      "enum": [
       "DR",
       "XR",
       "XSR",
       "SR",
       "LSR",
       "LR",
       "ER",
       "EER"
      ],
      "title": "Quality"
     },
     "enabled": {
      "type": "boolean",
      "title": "Enabled"
     },
     "created": {
      "type": "string",
      "format": "date-time",
      "title": "Created"
     }
    },
    "type": "object",
    "required": [
     "id",
     "query",
     "title_only",
     "quality",
     "enabled",
     "created"
    ],
    "title": "Rule"
   },
   "RuleCreate": {
    "properties": {
     "query": {
      "type": "string",
      "maxLength": 100,
      "minLength": 1,
      "title": "Query",
      "description": "matched case-insensitively (NFKC) against the title, or title + description"
     },
     "broadcasting": {
      "anyOf": [
       {
        "type": "string",
        "enum": [
         "td",
         "bs",
         "cs",
         "bs4k",
         "cs4k"
        ]
       },
       {
        "type": "null"
       }
      ],
      "title": "Broadcasting"
     },
     "service_id": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Service Id"
     },
     "title_only": {
      "type": "boolean",
      "title": "Title Only",
      "default": true
     },
     "quality": {
      "anyOf": [
       {
        "type": "string",
        "enum": [
         "DR",
         "XR",
         "XSR",
         "SR",
         "LSR",
         "LR",
         "ER",
         "EER"
        ]
       },
       {
        "type": "null"
       }
      ],
      "title": "Quality"
     }
    },
    "type": "object",
    "required": [
     "query"
    ],
    "title": "RuleCreate"
   },
   "RuleUpdate": {
    "properties": {
     "enabled": {
      "anyOf": [
       {
        "type": "boolean"
       },
       {
        "type": "null"
       }
      ],
      "title": "Enabled"
     },
     "quality": {
      "anyOf": [
       {
        "type": "string",
        "enum": [
         "DR",
         "XR",
         "XSR",
         "SR",
         "LSR",
         "LR",
         "ER",
         "EER"
        ]
       },
       {
        "type": "null"
       }
      ],
      "title": "Quality"
     },
     "title_only": {
      "anyOf": [
       {
        "type": "boolean"
       },
       {
        "type": "null"
       }
      ],
      "title": "Title Only"
     }
    },
    "type": "object",
    "title": "RuleUpdate"
   },
...
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     },
     "input": {
      "title": "Input"
     },
     "ctx": {
      "type": "object",
      "title": "Context"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   },
...
  },
  "securitySchemes": {
   "HTTPBearer": {
    "type": "http",
    "scheme": "bearer"
   }
  }
 }
}